requirements.txt
Django==4.2.7
djangorestframework==3.15.2 
numpy==2.0.2  
pandas==2.2.3
scikit-learn==1.5.2
spacy==3.8.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5 
joblib==1.4.2
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl
tqdm==4.67.1
